+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-(nothing yet!)
+### Fixed
+- Fix usage on module singleton classes
 
 ## [1.0.0] - 2021-06-24
 ### Added