+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fix usage on module singleton classes
+- Fix usage on module which would be extended by other classes
 
 ## [1.0.0] - 2021-06-24
 ### Added