+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
+### Updated
+- (Nothing, yet)
+### Fixed
+- (Nothing, yet)
+### Breaking Changes
+- None
+
+## [1.1.0] - 2021-07-29
 ### Updated
 - Improve performance across the board by:
   - removing `Hash#fetch`
@@ -16,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix usage on module singleton classes
 - Fix usage on module which would be extended by other classes
+### Breaking Changes
+- None
 
 ## [1.0.0] - 2021-06-24
 ### Added
@@ -70,7 +80,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Panolint
 - Dependabot setup
 
-[Unreleased]: https://github.com/panorama-ed/memo_wise/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/panorama-ed/memo_wise/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/panorama-ed/memo_wise/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/panorama-ed/memo_wise/compare/v0.4.0...v1.0.0
 [0.4.0]: https://github.com/panorama-ed/memo_wise/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/panorama-ed/memo_wise/compare/v0.2.0...v0.3.0